First commit of Tooz
For now Tooz is composed of a simple API and use ZooKeeper
as the first backend.

The current API provide the group membership protocol like
'servicegroup' in Nova, we enriched it by adding the notion
of capabilities. The API is currently synchronous, we plan to
design an asynchronous one. The notification mechanism is not yet
available. Please read the associated blueprint for more details:
https://wiki.openstack.org/wiki/Oslo/blueprints/service-sync

All new backends must implement api.CoordinationDriver and
validate the unit tests. Tooz will provide a unique unit-tests suite
which will be run against all backends.

Change-Id: I4e4cefb52cf6c39a3f6ba35ab1d4ef34a77b5811