Use the more reliable sysv_ipc instead of posix_ipc+lockutils
Remove the usage of lockutils (and surronding oslo modules)
and keep it much simpler by just using the sysv_ipc module
directly.

This adjustment reduces the complexity, test exposure and
API complexity of tooz; removing much of the complex functionality
that was pulled in when sucking in the lockutils module for its
simple (but still broken) posix_ipc logic.

Change-Id: I46db9bda252473317a0f3d65fe0329ac2151f332