Unlock the kazoo version
Latching on to an exact version is going to make it for consumers
of tooz+kazoo, which may themselves use a newer version, to be able
to correctly install tooz.

I'm also pretty sure there isn't any reason to bind to exactly 1.3.1
and allowing newer versions (which for now is up to 2.0) should be
fine.

Change-Id: I0dc810c2d9dbf5ff04cb5000b843ad4ac772fc43