Standardize on the same lock acquire method definition
It appears the the other drivers have a acquire that does not
take a timeout, but the zookeeper driver has a acquire method that
does take a timeout.

To be more consistent just standardize on the acquire not taking
a timeout paramater and having it provided by the driver options
like the IPC and memcached drivers do.

Change-Id: I86d5197ad57a4c891615e100365347e3e0bdd4c3