Split up the requirements for py2.x and py3.x
Certain requirements are not needed in py3.x since they
exist in the base python tree so it is good to split off
the requirements files and add the backported requirements
to the py2.x requirements and not the py3.x requirements.

The main current difference is the concurrent.futures
package which is built-in to python 3.x as part of the
standard library.

Change-Id: I7274cb597eb1e84d13d37f99508f1bdabe108d8c