Adjust the timeout to reflect the repeated retries
When a acquisition retry is triggered the timeout should
be adjusted to reflect the prior attempt instead of restarting
from its initial timeout value.

This makes it so that a acquire() request that ends up spinning
with a timeout of say 5 seconds will not repeatly spin for an
infinite number of 5 seconds, but instead will actually stop being
acquired after 5 seconds is reached.

Change-Id: Iff28ad972fdcec2784a868ae44beef5a2fa20b4e