Various fixes for locks and version compatibility
Avoid using lua locks

The default redis library will try to be smart about using
or not using lua based locks, it seems to not correctly predict
this correctly and then fails with errors such as:

''unknown command 'EVALSHA''

Upstream issue filed:

https://github.com/andymccurdy/redis-py/issues/550

The 'traditional' locks require PEXPIRE used by the redis-py
library which requires redis 2.6+ to actually work. So when
we are using a server version that is not new enough
we should conditionally raise a not implemented error to
avoid it later failing (which it will).

This also fixes a bunch of encoding issues that the enabling
of the py33 system now exposed (damn encodings...)

Change-Id: Ied6045276c9a1398db710a68aa75fe92aac9b4bb