Since we use msgpack this can be more than a str
It is often nice to have capabilities that are dicts
to allow for more complex capabilities that; and since
we use msgpack this is already enabled for usage (since
msgpack will serialize/deserialize capabilities provided
as dicts already).

This updates the zookeeper driver to ensure that capabilites
sent are first formatted to msgpack format (before being
sent) so that all drivers work in the same manner and accept
the same kind of items.

Change-Id: I2e19e806d38f70175706610cd92d9e22f606b41b