Allow coordinator non-string options and use them
To allow the zake driver to use a pre-existing storage
backend (for example, shared with a taskflow test that
also uses zake) we need a way to provide non-query-string
options and to get those options to the zake driver for
usage.

This approach allows the `get_coordinator` function
to take **kwargs and have those take precedence over
the options found with the same name in the urls query
string (and merge those together to form the coordinators
option dictionary that exists already).

Change-Id: Ibc745927eccac050992227410cca393d956f8c44