Remove support for redis < 2.6.0
To enable lua (and eval) usage to be accessible we
need to bump the min version to 2.6 (which should be
available on all distros that openstack is running
on).

This does that version bump and removes the code that
was handling and checking older versions (and the
differences/lack of APIs).

Change-Id: Ie8cd99d55e1907bb545b6692834b03871d9c1f83