Cap kazoo and zake from stable/juno global-requirements
The latest version of zake blocks kazoo 2.1 which causes version
conflicts in stable/juno jobs.

Cap kazoo and zake using the values from global-requirements in
stable/juno so that we can unblock stable/juno and stable/kilo (due to
grenade jobs on kilo).

Also updates .gitreview for stable/juno since this is the first
commit to stable/juno for tooz.

Partial-RTC: #1459322

Change-Id: Ib6f79cfcd18a758ce4b2a86f3ff323d533db4640