Raise exception on failed lock's CM acquire fail
In lock's context manager we are not checking that acquire returns true
before entering managed code.

Even though we don't except acquire to return False, we should still
confirm that the lock has been really acquired and we can proceed, and
raise an exception if it failed.

This patch modified metaclass Lock to raise new exception
LockAcquireFailed if it didn't acquire the lock.

Change-Id: I6d6f14ffd9ea10dd1ca203a92c987334e8a00416