Reduce duplication of code in handling multi-type blocking argument
The 'blocking' argument typically either means blocking as a bool
or blocking true but with the given timeout, so in various drivers they
handle this multi-type argument in nearly the same manner. To unify
that usage we can just have a utils method that breaks the multi-type
argument into its components and uniformly use that in drivers.

Change-Id: I0a6c4b50919fab72d383b9c9f2e772fc931cb6b8