file: make python2 payload readable from python3
We ensure that everything written in file are explicitly bytes or unicode.
This allows to be able to read file written by python2 with python3.

This also needs to use new oslo.serialization that fix the same
kind of issue for the datetime object.

This can be manually tested by running:

   rm -rf /tmp/groups ; tox -epy27,py34

Change-Id: I46988e3f3cb738f6a709367f851c1fac79446ede
Closes-bug: #1556081