etcd: don't run heartbeat() concurrently
As the bug specifies, running heartbeat() concurrently to other actions
(such as release()) is raceful as using heartbeat() can and will change
the modifiedIndex field of a lock, thus breaking the compare function of
the Compare-and-Delete used by release(). This patch, therefore, uses
threading's Lock object to make sure only a single action is done
concurrently in tooz' etcd driver.

Closes-Bug: #1603005
Change-Id: Ief32a3d5add90ebbd4adf825d78fca3dd57e5d9e