Rewrite heartbeat runner with event
The current implementation uses an event attached to the Heart class to detect
if the thread is done. Since the thread is a daemon, it might be killed after
the interpreter is done, especially if an application is killed before calling
stop() on the coordinator. This results in the following backtrace:

Exception in thread Thread-1 (most likely raised during interpreter shutdown):
Traceback (most recent call last):
  File "/usr/lib/python2.7/threading.py", line 801, in __bootstrap_inner
  File "/usr/lib/python2.7/threading.py", line 754, in run
  File "/usr/local/lib/python2.7/dist-packages/oslo_utils/excutils.py", line 250, in wrapper
  File "/usr/local/lib/python2.7/dist-packages/tooz/coordination.py", line 206, in _beat_forever_until_stopped
  File "/usr/lib/python2.7/threading.py", line 585, in set
  File "/usr/lib/python2.7/threading.py", line 407, in notifyAll
<type 'exceptions.TypeError'>: 'NoneType' object is not callable

This patches simplifies the heart class by using the thread status rather than
an event to check if it's alive and wait for it to stop.

Change-Id: I03ccbefd3d68e9dd85ea23a3840d606d1ffc80bb
(cherry picked from commit 325324137649cc40791260b6d12df33c0407e039)
(cherry picked from commit bc11f09b831dd5811d42f4845117aea2581c194e)