FileDriver:Support multiple processes
The interprocess file-lock is not really released when user call
FileLock.release() method. So the second process will be blocked
when it try to acquire a lock which has been 'released' by the
first process until the first process is exited. This lead to
current filelock don't support multiple processes. This patch
fixed this problem.

Closes-Bug: #1665897
Change-Id: I3d4331a400dd551fa97cada2d7af9415fda720da