etcd: fix acquire(blocking=True) on request exception
If requests raises an exception and blocking is True, then there's no watch and
nothing is returned. The problem is that then the reply variable is checked but
it is actually not assigned, so a NameError is raised.

This makes sure False is returned if an error occurs and blocking is True.

Change-Id: I078826a894ee2cb8754259c8db9ea5ae19114910