import zuul job settings from project-config
This is a mechanically generated patch to complete step 1 of moving
the zuul job settings out of project-config and into each project
repository.

Because there will be a separate patch on each branch, the branch
specifiers for branch-specific jobs have been removed.

See the python3-first goal document for details:
https://governance.openstack.org/tc/goals/stein/python3-first.html

Depends-On: I8247a51c26dea4e8b2834b9655ffbb37b36a6af4
Change-Id: I7e9a3124a824d00623fdbb44ece9456b0901788a
Story: #2003250