Switch to autodoc_default_options
autodoc_default_flags was deprecated and is now triggering a warning
that breaks our builds. Switch to the equivalent autodoc_default_options
configuration, which essentially turns the flags into dict keys that
map to None[1]: 'Setting None is equivalent to giving the option name
in the list format (i.e. it means “yes/true/on”).'

Change-Id: I573f19a054e535a90db8bb16a02b2c7f28d8e6d1
1: http://www.sphinx-doc.org/en/master/usage/extensions/autodoc.html#confval-autodoc_default_options