Unblock tooz gate
We have a couple of issues blocking tooz test jobs right now. This
is a combination of two commits that fix the problem.

Cap grpcio<1.16.0

1.16.0 introduces a bug that is causing our etcd3 tests to fail on
a consistent basis. Until that is resolved we can't use any higher
versions of grpcio.

1.0.0 is used as the lower bound on the assumption that any version
of the current major release will work. Since we weren't setting a
bound before that seems relatively safe.

Partial-Bug: 1808046

Fix requirements

The requirements check job is failing on a couple of the entries in
this project. enum34 needs a lower bound, and futurist has an
exclusion that isn't in global-requirements.

enum34 is given the same lower bound as oslo.upgradecheck, and
futurist's lower bound is increased to the last global lower bound
(before lower bounds were removed), which eliminates the need for
the exclusion.

Change-Id: I1b279e138325b596f5ec2c61717029d2f7bdd075
Depends-On: https://review.openstack.org/#/c/624759/