Avoid redis lock's expire_time exceeding timeout.
After n heartbeats, the expireation time of the lock key becomes
(1 + 0.5n) * timeout! Therefore, when the lock holder performs a
long task and exits abnormally, the lock will not be automatically
released in time.

Changes:
  setup.cfg
  tooz/drivers/redis.py

NOTE(stephenfin): We have to remove the changes to requirements in
setup.cfg since it's not possible to raise the minimum version of a
requirement on a stable branch. Instead we check to see if the attribute
added in the new version is present and use it only if so.

Change-Id: I2437071bb575e95083b0d57a6a0a33aaf8356113
(cherry picked from commit 20bda6ed999e71823b6eac90f01d1a29898b63b2)