Unblock tooz gate
This is a squash of two commits needed to get gate tests passing
again on tooz. See below for details:

Move testing to py36

The base infra images no longer have python 3.5 available and are
all failing. We should be testing on py36 now anyway since that is
the minimum supported version for Train.

Stop redis-server before running tests

Just installing redis-server on Ubuntu Bionic starts the service,
which means when we try to start one via pifpaf it fails due to the
port already being in use.

This change adds a pre-run playbook that stops the redis-server
service so ours can run successfully.

Change-Id: I3f7f7ea9069d8c890a82f31ad14c9663e98c09dc
Closes-Bug: 1828610