Deprecate the etcd3 driver
The etcd3 driver depends on the python-etcd3 [1]
library that is currently stale and no has no
active maintenance [2] with a non-working CI
that blocks all PRs.

It's also currently blocking Tooz from supporting
newer versions of tenacity [3] that a lot of
projects is now pinning because newer version of
tenacity dropped old legacy compatibility code
that has been in there for years.

We currently have etcd3gw as an alternative driver
where the library the driver depends on is a
deliverable that we maintain.

[1] https://github.com/kragniz/python-etcd3
[2] https://github.com/kragniz/python-etcd3/issues/1719
[3] https://review.opendev.org/c/openstack/tooz/+/829412

Change-Id: Iff0cd3b51cfc45ecbddc831c55267e80e9c79cac