Remove unnecessary dependencies
pbr is not a runtime dependency, so there's no need to list it as one.
Similarly, we don't need to install pre-commit in our test environment
nor do we use nose of python-subunit (at least directly). Remove them
all.

We also remove so notes about dependency order that have not been
relevant since the new dependency resolver was introduced in pip 20.3.

Change-Id: Iea7a68dae4c96bc7118e946fc6d494a8799942af
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>